Improve name of new console log function
The new console log function here isn't any more of a channel log than
the old one.  The name was bad, and no one could have been using this
yet since we just ran reswig.  We'll take the same tack the kernel
developers do when replacing a function with an otherwise equivalent
one with a better API.